Updates to the Docker configs
- Implements docker-secrets and adds the secrets shim.
- Updates the Dockerfile to use our standard 3-stage build.
- Refactors the configs. All are now defined in the application.rb,
which you can check to see what is needed elsewhere in the app. The
EXPECT_URL is a URI that includes the scheme (ssh), user, host, and path
to the expect script run to update the patron record. See the default
value in application.rb for its format.
- The PRIVATE_KEY variable contains the actual contents of the private
key used to execute the expect script over SSH. It's no longer just the
path to the file. There's room for improvement here; for instance, we
could dispense with this entirely by simply copying that file to
~/.ssh/id_rsa inside the container.