Add UCOP request form and re-implement AltMedia
- Adds the UCOP request form, building on changes that @ghill had made.
- Converts the UCOP form and the existing AltMedia form to use form
objects. This is critical to incorporating more forms down the road (we
have two new ones to integrate plus old ones.)
- Adds an unauthenticated home/landing page that links to the available
forms. The list is created manually, for now, but eventually will be
populated via a database or metaprogramming.
- Standardizes the view templates using Bootstrap. Moves the
login/logout links, which had been scattered throughout different
templates, into a new navbar template.
- Adds an :authenticate! method, intended as a before_action filter,
which forces the user to login. Actions should perform additional
validation as needed. The login flow now populates session[:user] with
(some) of the data returned by CalNet. This data can be used to
instantiate a User object, which in turn can be used to access the
associated patron record's data.
- Adds a bunch of tests...