Adds staging environment
- Adds config/environments/staging.rb, which is a clone of production
except for the addition of a StagingMailInterceptor, which routes all
outgoing emails to lib-testmail@lists.berkeley.edu. to/cc/bcc are
preserved in headers named X-Original-{To,CC,BCC}.
- Splits configs into a base docker-compose plus overrides for each
environment (dev, the default, plus ci, staging, and production).
- Makes Jenkins auto-deploy "latest" to staging. Production is pinned to
a specific tag, so to deploy a version you need to add a new commit
modifying the config.