Adds top-level error handlers
- Creates the Framework::Errors module for application-specific
exceptions. The first set includes a base FrameworkError and two
sub-classes: NotAuthenticatedError, for triggering the authentication
flow; and PatronApiError, for wrapping errors that occur when hitting
the Patron API.
- Adds an ErrorHandling controller concern, which includes our various
rescue_from directives. The directives should either issue a redirect
(as for NotAuthenticatedError, which redirects to the login page) or
render and error view (app/views/errors/...) with the appropriate HTTP
status code.