Queue jobs with patronID instead of partial record
Having jobs lookup patrons by ID ensures that they have a full and
up-to-date patron record before making changes. This is particularly
important for notes, which currently are not passed to the jobs. By
having jobs lookup the patron, we can more easily implement a check to
avoid adding duplicate notes.