Implements /health endpoint
Adds a healthcheck which verifies that we can lookup patrons via the
Patron API. If that check succeeds, returns 200; otherwise returns 429
(Too Many Requests--this weirdness is per the Consul spec). We return
429 instead of 500 in order to distinguish between "warnings" (4xx) and
genuine failure (5xx); the application is unhealthy in either case, but
with a 5xx you'd expect to need to restart the instance.