Ignore CVE-2015-9284 when running bundle:audit
We opted to change how the bundle:audit task worked rather than write a
convoluted shell script in Jenkinsfile. After inspecting
Bundle::Audit::Task, we determined it called Bundle::Audit::CLI, and
that behind the scenes this was calling ::Thor.start. This method just
takes an array of arguments as you would've passed them on the command
line, so we added the appropriate --ignore flag to the "check" command.

Links:

- Thor: https://github.com/erikhuda/thor/wiki/Getting-Started
- Bundle::Audit::Task: https://github.com/rubysec/bundler-audit/blob/master/lib/bundler/audit/task.rb