Jenkins: Avoid credentials for portainer webhooks
Portainer webhooks don't need to be secrets, for a few reasons:
- Portainer's ingress configuration blocks web requests from outside of
trusted ranges.
- The webhook is only capable of triggering a service update, which in
almost every case is a no-op. (Docker will check for an update and,
seeing that there is none, do nothing.) The worst-case scenario is that
Docker issues a lot of checksum-match requests to the registry, which is
not a big deal.
- Using Jenkins' credentials complicates making updates to the pipeline
configuration. Where possible, we should avoid them, either offloading
that into another system or leveraging globally-available credentials
which do not need to be customized per-job.