Pipeline: Warn (don't fail) on security issues
- Security warnings reported by bundle:audit or brakeman now mark the
stage as "unstable" but do not fail the build.
- GitLab treats unstable builds as failures. As a workaround, we use the
verbose catchError() pipeline method rather than the more concise and
obvious warnError(), because the former allows us to mark the stage as
unstable while retaining the previous status for the overall build.
- Relatedly, GitLab will not show that you have vulnerabilities. You
must click through to the Jenkins job to see that.