AP-184: Bumps SMTP timeouts to avoid double-emails
Increases SMTP open and read timeouts to prevent double-emails when
sending emails with large attachments. See the linked Slack thread for
details on how this was debugged, but in a nutshell, sending a large
attachment can cause Rails to timeout the email, send the failure email,
then have the success email succeed after all in the background. This is
a consequence of Ruby's well-known Timeout traps.

@see https://ucb-lit.slack.com/archives/CC56XDDT2/p1692726040059019
@see https://www.exceptionalcreatures.com/bestiary/Timeout/Error