Rework Stack Pass administrator name handling
Instead of having the name edited by site admins, we update it on the
user's log in action from CAS.  This ensures it is always up to date.

This also adds a method for the human-friendly status of the request.
This helps DRY the views.

This additionally adds a column to the database for the FrameworkUser
ID of who processed the request, so that the updates cascade.