Rework Stack Pass administrator name handling
Instead of having the name edited by site admins, we update it on the
user's log in action from CAS.  This ensures it is always up to date.

This additionally adds a column to the database for the FrameworkUser
ID of who processed the request, so that the updates cascade.

TBD: Migration logic should be added that will search the database for
existing names and replace them with the FrameworkUser ID, so that
updates cascade for these users as well.